# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JobPilot Agent",
    "description": "JobPilot AI Agent 服务层，提供 JD 路由分析与模拟面试功能。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/skills/{name}/dispatch-check": {
      "post": {
        "tags": [
          "Skills Debug"
        ],
        "summary": "检查指定 Skill 是否会被触发",
        "description": "给定 JDContext 片段，调用指定 Skill 的 should_invoke() 并返回结果。用于验证路由规则是否符合预期，不会真正执行 Skill。",
        "operationId": "dispatch_check_api_v1_skills__name__dispatch_check_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/{name}/invoke": {
      "post": {
        "tags": [
          "Skills Debug"
        ],
        "summary": "直接调用指定 Skill",
        "description": "直接触发某个 Skill 执行，绕过 LangGraph 流程。用于人工调试和验收测试，返回 SkillOutput 原始 JSON。",
        "operationId": "invoke_skill_api_v1_skills__name__invoke_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvokeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DispatchCheckRequest": {
        "properties": {
          "jd_text": {
            "type": "string",
            "title": "Jd Text",
            "default": "示例 JD 文本"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "debug_user"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": "debug_req"
          },
          "user_context": {
            "$ref": "#/components/schemas/UserContext",
            "default": {
              "preferred_lang": "zh"
            }
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JDClassification"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "DispatchCheckRequest",
        "description": "dispatch-check 端点的请求体，携带足够的 JDContext 字段用于路由判断。"
      },
      "DispatchCheckResponse": {
        "properties": {
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "should_invoke": {
            "type": "boolean",
            "title": "Should Invoke"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "skill_name",
          "should_invoke"
        ],
        "title": "DispatchCheckResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvokeRequest": {
        "properties": {
          "jd_text": {
            "type": "string",
            "title": "Jd Text"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "debug_user"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": "debug_req"
          },
          "user_context": {
            "$ref": "#/components/schemas/UserContext",
            "default": {
              "preferred_lang": "zh"
            }
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JDClassification"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeout_seconds": {
            "type": "integer",
            "title": "Timeout Seconds",
            "default": 15
          }
        },
        "type": "object",
        "required": [
          "jd_text"
        ],
        "title": "InvokeRequest",
        "description": "invoke 端点的请求体。"
      },
      "JDClassification": {
        "properties": {
          "job_type": {
            "type": "string",
            "enum": [
              "tech",
              "product",
              "design",
              "ops",
              "mgmt"
            ],
            "title": "Job Type"
          },
          "sub_type": {
            "type": "string",
            "title": "Sub Type"
          },
          "level": {
            "type": "string",
            "enum": [
              "junior",
              "middle",
              "senior",
              "lead"
            ],
            "title": "Level"
          },
          "locale": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Locale"
          },
          "channel": {
            "type": "string",
            "enum": [
              "social",
              "campus"
            ],
            "title": "Channel"
          }
        },
        "type": "object",
        "required": [
          "job_type",
          "sub_type",
          "level",
          "locale",
          "channel"
        ],
        "title": "JDClassification"
      },
...
      "UserContext": {
        "properties": {
          "user_kb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Kb Id"
          },
          "preferred_lang": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Preferred Lang",
            "default": "zh"
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          },
          "portfolio_doc_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Doc Ref"
          }
        },
        "type": "object",
        "title": "UserContext"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}